moved ExecutionStatusPopover to a shared place
