+++ app/src/pages/inside/manualLaunchesPage/executionStatusPopover/messages.ts
